Major Changes to Gate implementation
Replaced gate functions in gates.py with returns of the gate matricies directly
Added gate functions for zeroQubit and oneQubit (|0> and |1>)
Added a new function, apply_gate(), which takes a gate matrix, the qubit matrix, the position of the qubit to modify, and (optionally) a list of control qubit positions. It uses the other new functions make_full_gate() and make_complete_identity_tensor() to apply the gate correctly, with all controls applied, as it would occur over the entire qubit matrix. Untested, but should work with qubit matricies of any size, with any number of controls, and with any basic 2x2 gate.
Modified the methods in Simulator.ipynb to reflect the new gates.py